backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
google-cloud-vision==3.7.2
pillow==10.2.0
python-dotenv==1.0.1
